Python/requirements.txt
google-api-python-client>=1.7.4
pytz>=2018.6
notify-run>=0.0.10
